mod. ArticleRepository.class JPA로 수정
ArticleControllerTests.class에서 2개 작동 안함
