+++ src/main/java/techcourse/myblog/domain/Article.java
@@ -1,12 +1,23 @@
 package techcourse.myblog.domain;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.util.Objects;
 
+@Entity
 public class Article {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private String title;
     private String coverUrl;
     private String contents;
+
+    public Article() {
+
+    }
 
     public Article(long id, String title, String coverUrl, String contents) {
         this.id = id;
@@ -40,6 +51,22 @@ public class Article {
 
     public String getContents() {
         return contents;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public void setCoverUrl(String coverUrl) {
+        this.coverUrl = coverUrl;
+    }
+
+    public void setContents(String contents) {
+        this.contents = contents;
     }
 
     @Override
