+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -5,14 +5,11 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
-import techcourse.myblog.domain.ArticleDto;
+import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.ArticleRepository;
-
-import java.util.List;
 
 @Controller
 public class ArticleController {
-
     private final ArticleRepository articleRepository;
 
     @Autowired
@@ -22,7 +19,7 @@ public class ArticleController {
 
     @GetMapping("/")
     public String index(Model model) {
-        List<ArticleDto> articles = articleRepository.findAll();
+        Iterable<Article> articles = articleRepository.findAll();
         model.addAttribute("articles", articles);
         return "index";
     }
@@ -33,33 +30,37 @@ public class ArticleController {
     }
 
     @PostMapping("/articles")
-    public RedirectView addArticle(final ArticleDto articleParam) {
-        long latestId = articleRepository.add(articleParam);
-        return new RedirectView("/articles/" + latestId);
+    public RedirectView addArticle(final Article articleParam) {
+        Article article = articleRepository.save(articleParam);
+        return new RedirectView("/articles/" + article.getId());
     }
 
     @GetMapping("/articles/{articleId}")
     public String showArticleById(@PathVariable final long articleId, final Model model) {
-        ArticleDto article = articleRepository.findById(articleId);
+        Article article = articleRepository.findById(articleId).orElseThrow(IllegalArgumentException::new);
         model.addAttribute("article", article);
         return "article";
     }
 
     @PutMapping("/articles/{articleId}")
-    public RedirectView updateArticle(@PathVariable final long articleId, final ArticleDto articleParam) {
-        long updateId = articleRepository.updateById(articleParam, articleId);
-        return new RedirectView("/articles/" + updateId);
+    public RedirectView updateArticle(@PathVariable final long articleId, final Article articleParam) {
+        articleRepository.update(articleParam);
+        articleParam.setId(articleId);
+        Article article = articleRepository.findById(articleId).orElseThrow(IllegalArgumentException::new);
+        return new RedirectView("/articles/" + article.getId());
     }
 
     @DeleteMapping("articles/{articleId}")
     public RedirectView deleteArticle(@PathVariable final long articleId) {
+        System.out.println(articleId);
+        System.out.println(articleRepository.findAll());
         articleRepository.deleteById(articleId);
         return new RedirectView("/");
     }
 
     @GetMapping("/articles/{articleId}/edit")
     public String updateArticle(@PathVariable final long articleId, final Model model) {
-        ArticleDto article = articleRepository.findById(articleId);
+        Article article = articleRepository.findById(articleId).orElseThrow(IllegalArgumentException::new);
         model.addAttribute("article", article);
         return "article-edit";
     }
